Refactoring; Update Box Detection;

+++ src/main/kotlin/me/pseudoapp/other/CommonUtils.kt
@@ -62,4 +62,4 @@ object SizeObserver {
     }
 }
 
-fun Float.convertToPx(): Int = (this / SizeObserver.oneDpValue).toInt()
+fun Float.convertToPx(): Float = (this / SizeObserver.oneDpValue)
